requirements.txt
openai-whisper>=20231117
streamlit>=1.27.0
torch>=2.0.0
torchaudio>=2.0.0
numpy
pandas>=1.3.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1
